reverting -r1822816
The main change of the original commit was to move the TCCL resolution and calling setExtension.
This seems to break the Tomcat ParallelWebAppClassLoader.
Rather it seems that the cleanup then is being performed on the tomcat side, even if the server is already shut down.

I've reverted this change for now. up 4 review!


git-svn-id: https://svn.apache.org/repos/asf/openwebbeans/meecrowave/trunk@1824558 13f79535-47bb-0310-9956-ffa450edef68