MEECROWAVE-337 move from javax JavaEE8 to jakarta EE 10
This is the native support for Jakarta EE 10.
Note that a few libs need release, e.g. CXF 4.1.0 as they did use a deprecated CDI API
which got now removed.

Further note that oauth2 is right now in a broken shape. We need to migrate
the keystore handling over from com.sun.* to bouncycastle.
I've disabled the corresponding tests for now.